+++ certifications.go
@@ -2,18 +2,6 @@ package tmdb
 
 import "fmt"
 
-// Certification type is a struct for a single certification JSON response.
-type Certification struct {
-	Certification string `json:"certification"`
-	Meaning       string `json:"meaning"`
-	Order         int    `json:"order"`
-}
-
-// Certifications type is a struct for movie and tv certifications JSON response.
-type Certifications struct {
-	Certifications map[string][]Certification `json:"certifications"`
-}
-
 // GetCertificationMovie get an up to date list of the
 // officially supported movie certifications on TMDb.
 //
