+++ changes.go
@@ -2,12 +2,6 @@ package tmdb
 
 import "fmt"
 
-// ChangesMovie type is a struct for movie changes JSON response.
-type ChangesMovie struct {
-	*ChangesMovieResults
-	PaginatedResultsMeta
-}
-
 // GetChangesMovie get a list of all of the movie ids
 // that have been changed in the past 24 hours.
 //
@@ -15,7 +9,7 @@ type ChangesMovie struct {
 // at a time with the start_date and end_date query parameters.
 // 100 items are returned per page.
 //
-// https://developers.themoviedb.org/3/changes/get-movie-change-list
+// https://developer.themoviedb.org/reference/changes-movie-list
 func (c *Client) GetChangesMovie(
 	urlOptions map[string]string,
 ) (*ChangesMovie, error) {
@@ -34,42 +28,6 @@ func (c *Client) GetChangesMovie(
 	return &changesMovies, nil
 }
 
-// ChangesTV type is a struct for tv changes JSON response.
-type ChangesTV struct {
-	*ChangesMovie
-}
-
-// GetChangesTV get a list of all of the TV show ids
-// that have been changed in the past 24 hours.
-//
-// You can query it for up to 14 days worth of changed IDs
-// at a time with the start_date and end_date query parameters.
-// 100 items are returned per page.
-//
-// https://developers.themoviedb.org/3/changes/get-tv-change-list
-func (c *Client) GetChangesTV(
-	urlOptions map[string]string,
-) (*ChangesTV, error) {
-	options := c.fmtOptions(urlOptions)
-	tmdbURL := fmt.Sprintf(
-		"%s%schanges?api_key=%s%s",
-		baseURL,
-		tvURL,
-		c.apiKey,
-		options,
-	)
-	changesTV := ChangesTV{}
-	if err := c.get(tmdbURL, &changesTV); err != nil {
-		return nil, err
-	}
-	return &changesTV, nil
-}
-
-// ChangesPerson type is a struct for person changes JSON response.
-type ChangesPerson struct {
-	*ChangesMovie
-}
-
 // GetChangesPerson get a list of all of the person ids
 // that have been changed in the past 24 hours.
 //
@@ -77,7 +35,7 @@ type ChangesPerson struct {
 // at a time with the start_date and end_date query parameters.
 // 100 items are returned per page.
 //
-// https://developers.themoviedb.org/3/changes/get-person-change-list
+// https://developer.themoviedb.org/reference/changes-people-list
 func (c *Client) GetChangesPerson(
 	urlOptions map[string]string,
 ) (*ChangesPerson, error) {
@@ -94,4 +52,30 @@ func (c *Client) GetChangesPerson(
 		return nil, err
 	}
 	return &changesPerson, nil
+}
+
+// GetChangesTV get a list of all of the TV show ids
+// that have been changed in the past 24 hours.
+//
+// You can query it for up to 14 days worth of changed IDs
+// at a time with the start_date and end_date query parameters.
+// 100 items are returned per page.
+//
+// https://developer.themoviedb.org/reference/changes-tv-list
+func (c *Client) GetChangesTV(
+	urlOptions map[string]string,
+) (*ChangesTV, error) {
+	options := c.fmtOptions(urlOptions)
+	tmdbURL := fmt.Sprintf(
+		"%s%schanges?api_key=%s%s",
+		baseURL,
+		tvURL,
+		c.apiKey,
+		options,
+	)
+	changesTV := ChangesTV{}
+	if err := c.get(tmdbURL, &changesTV); err != nil {
+		return nil, err
+	}
+	return &changesTV, nil
 }
