+++ collections.go
@@ -2,34 +2,9 @@ package tmdb
 
 import "fmt"
 
-// CollectionDetails type is a struct for details JSON response.
-type CollectionDetails struct {
-	ID           int64  `json:"id"`
-	Name         string `json:"name"`
-	Overview     string `json:"overview"`
-	PosterPath   string `json:"poster_path"`
-	BackdropPath string `json:"backdrop_path"`
-	Parts        []struct {
-		Adult            bool    `json:"adult"`
-		BackdropPath     string  `json:"backdrop_path"`
-		GenreIDs         []int64 `json:"genre_ids"`
-		ID               int64   `json:"id"`
-		MediaType        string  `json:"media_type"`
-		OriginalLanguage string  `json:"original_language"`
-		OriginalTitle    string  `json:"original_title"`
-		Overview         string  `json:"overview"`
-		PosterPath       string  `json:"poster_path"`
-		ReleaseDate      string  `json:"release_date"`
-		Title            string  `json:"title"`
-		Video            bool    `json:"video"`
-		Popularity       float32 `json:"popularity"`
-		VoteMetrics
-	} `json:"parts"`
-}
-
 // GetCollectionDetails get collection details by id.
 //
-// https://developers.themoviedb.org/3/collections/get-collection-details
+// https://developer.themoviedb.org/reference/collection-details
 func (c *Client) GetCollectionDetails(
 	id int,
 	urlOptions map[string]string,
@@ -46,22 +21,9 @@ func (c *Client) GetCollectionDetails(
 	return &collectionDetails, nil
 }
 
-// CollectionImage type is a struct for a single image.
-type CollectionImage struct {
-	ImageBase
-	Iso639_1 string `json:"iso_639_1"`
-}
-
-// CollectionImages type is a struct for images JSON response.
-type CollectionImages struct {
-	ID        int64             `json:"id"`
-	Backdrops []CollectionImage `json:"backdrops"`
-	Posters   []CollectionImage `json:"posters"`
-}
-
 // GetCollectionImages get the images for a collection by id.
 //
-// https://developers.themoviedb.org/3/collections/get-collection-images
+// https://developer.themoviedb.org/reference/collection-images
 func (c *Client) GetCollectionImages(
 	id int,
 	urlOptions map[string]string,
@@ -78,15 +40,9 @@ func (c *Client) GetCollectionImages(
 	return &collectionImages, nil
 }
 
-// CollectionTranslations type is a struct for translations JSON response.
-type CollectionTranslations struct {
-	ID           int64         `json:"id"`
-	Translations []Translation `json:"translations"`
-}
-
 // GetCollectionTranslations get the list translations for a collection by id.
 //
-// https://developers.themoviedb.org/3/collections/get-collection-translations
+// https://developer.themoviedb.org/reference/collection-translations
 func (c *Client) GetCollectionTranslations(
 	id int,
 	urlOptions map[string]string,
