+++ companies.go
@@ -4,25 +4,9 @@ import (
 	"fmt"
 )
 
-// CompanyDetails type is a struct for details JSON response.
-type CompanyDetails struct {
-	Description   string `json:"description"`
-	Headquarters  string `json:"headquarters"`
-	Homepage      string `json:"homepage"`
-	ID            int64  `json:"id"`
-	LogoPath      string `json:"logo_path"`
-	Name          string `json:"name"`
-	OriginCountry string `json:"origin_country"`
-	ParentCompany struct {
-		Name     string `json:"name"`
-		ID       int64  `json:"id"`
-		LogoPath string `json:"logo_path"`
-	} `json:"parent_company"`
-}
-
 // GetCompanyDetails get a companies details by id.
 //
-// https://developers.themoviedb.org/3/companies/get-company-details
+// https://developer.themoviedb.org/reference/company-details
 func (c *Client) GetCompanyDetails(
 	id int,
 ) (*CompanyDetails, error) {
@@ -40,16 +24,9 @@ func (c *Client) GetCompanyDetails(
 	return &companyDetails, nil
 }
 
-// CompanyAlternativeNames type is a struct for alternative
-// names JSON response.
-type CompanyAlternativeNames struct {
-	ID int64 `json:"id"`
-	*CompanyAlternativeNamesResult
-}
-
 // GetCompanyAlternativeNames get the alternative names of a company.
 //
-// https://developers.themoviedb.org/3/companies/get-company-alternative-names
+// https://developer.themoviedb.org/reference/company-alternative-names
 func (c *Client) GetCompanyAlternativeNames(
 	id int,
 ) (*CompanyAlternativeNames, error) {
@@ -67,19 +44,6 @@ func (c *Client) GetCompanyAlternativeNames(
 	return &companyAlternativeNames, nil
 }
 
-// CompanyImage type is a struct for a single image.
-type CompanyImage struct {
-	ImageBase
-	ID       string `json:"id"`
-	FileType string `json:"file_type"`
-}
-
-// CompanyImages type is a struct for images JSON response.
-type CompanyImages struct {
-	ID    int64          `json:"id"`
-	Logos []CompanyImage `json:"logos"`
-}
-
 // GetCompanyImages get a companies logos by id.
 //
 // There are two image formats that are supported for companies,
@@ -90,7 +54,7 @@ type CompanyImages struct {
 // An SVG can be scaled properly beyond those dimensions if you
 // call them as a PNG.
 //
-// https://developers.themoviedb.org/3/companies/get-company-images
+// https://developer.themoviedb.org/reference/company-images
 func (c *Client) GetCompanyImages(
 	id int,
 ) (*CompanyImages, error) {
