+++ discover.go
@@ -2,12 +2,6 @@ package tmdb
 
 import "fmt"
 
-// DiscoverMovie type is a struct for movie JSON response.
-type DiscoverMovie struct {
-	PaginatedResultsMeta
-	*DiscoverMovieResults
-}
-
 // GetDiscoverMovie discover movies by different types of data like
 // average rating, number of votes, genres and certifications. You can
 // get a valid list of certifications from the  method.
@@ -31,7 +25,7 @@ type DiscoverMovie struct {
 // separated. Comma's are treated like an AND and query while pipe's
 // are an OR.
 //
-// https://developers.themoviedb.org/3/discover/movie-discover
+// https://developer.themoviedb.org/reference/discover-movie
 func (c *Client) GetDiscoverMovie(
 	urlOptions map[string]string,
 ) (*DiscoverMovie, error) {
@@ -50,12 +44,6 @@ func (c *Client) GetDiscoverMovie(
 	return &discoverMovie, nil
 }
 
-// DiscoverTV type is a struct for tv JSON response.
-type DiscoverTV struct {
-	PaginatedResultsMeta
-	*DiscoverTVResults
-}
-
 // GetDiscoverTV Discover TV shows by different types of data like average
 // rating, number of votes, genres, the network they aired on and air dates.
 //
@@ -65,7 +53,7 @@ type DiscoverTV struct {
 // Also note that a number of filters support being comma (,) or pipe (|)
 // separated. Comma's are treated like an AND and query while pipe's are an OR.
 //
-// https://developers.themoviedb.org/3/discover/tv-discover
+// https://developer.themoviedb.org/reference/discover-tv
 func (c *Client) GetDiscoverTV(
 	urlOptions map[string]string,
 ) (*DiscoverTV, error) {
