+++ discover_struct.go
@@ -1,0 +1,13 @@
+package tmdb
+
+// DiscoverMovie type is a struct for movie JSON response.
+type DiscoverMovie struct {
+	PaginatedResultsMeta
+	*DiscoverMovieResults
+}
+
+// DiscoverTV type is a struct for tv JSON response.
+type DiscoverTV struct {
+	PaginatedResultsMeta
+	*DiscoverTVResults
+}
