+++ genres_struct.go
@@ -1,0 +1,6 @@
+package tmdb
+
+// GenreMovieList type is a struct for genres movie list JSON response.
+type GenreMovieList struct {
+	Genres []Genre `json:"genres"`
+}
