+++ keywords.go
@@ -2,15 +2,9 @@ package tmdb
 
 import "fmt"
 
-// KeywordDetails type is a struct for keyword JSON response.
-type KeywordDetails struct {
-	ID   int64  `json:"id"`
-	Name string `json:"name"`
-}
-
 // GetKeywordDetails get keyword details by id.
 //
-// https://developers.themoviedb.org/3/keywords/get-keyword-details
+// https://developer.themoviedb.org/reference/keyword-details
 func (c *Client) GetKeywordDetails(
 	id int,
 ) (*KeywordDetails, error) {
@@ -28,33 +22,16 @@ func (c *Client) GetKeywordDetails(
 	return &keywordDetails, nil
 }
 
-// KeywordMovies type is a struct for movies that belong to a keyword JSON response.
-type KeywordMovies struct {
-	ID      int64 `json:"id"`
-	Results []struct {
-		Adult            bool    `json:"adult"`
-		BackdropPath     string  `json:"backdrop_path"`
-		GenreIDs         []int64 `json:"genre_ids"`
-		ID               int64   `json:"id"`
-		OriginalLanguage string  `json:"original_language"`
-		OriginalTitle    string  `json:"original_title"`
-		Overview         string  `json:"overview"`
-		PosterPath       string  `json:"poster_path"`
-		ReleaseDate      string  `json:"release_date"`
-		Title            string  `json:"title"`
-		Video            bool    `json:"video"`
-		Popularity       float32 `json:"popularity"`
-		VoteMetrics
-	} `json:"results"`
-	PaginatedResultsMeta
-}
-
+// Deprecated: Use GetDiscoverMovie instead.
+//
 // GetKeywordMovies get the movies that belong to a keyword.
 //
 // We highly recommend using movie discover instead of this
 // method as it is much more flexible.
 //
-// https://developers.themoviedb.org/3/keywords/get-movies-by-keyword
+// This endpoint is deprecated in the TMDB API.
+//
+// https://developer.themoviedb.org/reference/keyword-movies
 func (c *Client) GetKeywordMovies(
 	id int,
 	urlOptions map[string]string,
