+++ movies.go
@@ -3,135 +3,11 @@ package tmdb
 import (
 	"fmt"
 	"net/http"
-
-	json "github.com/goccy/go-json"
 )
-
-// MovieDetails type is a struct for movie details JSON response.
-type MovieDetails struct {
-	Adult               bool                `json:"adult"`
-	BackdropPath        string              `json:"backdrop_path"`
-	BelongsToCollection BelongsToCollection `json:"belongs_to_collection"`
-	Budget              int64               `json:"budget"`
-	Genres              []Genre             `json:"genres"`
-	Homepage            string              `json:"homepage"`
-	ID                  int64               `json:"id"`
-	IMDbID              string              `json:"imdb_id"`
-	OriginalLanguage    string              `json:"original_language"`
-	OriginalTitle       string              `json:"original_title"`
-	Overview            string              `json:"overview"`
-	Popularity          float32             `json:"popularity"`
-	PosterPath          string              `json:"poster_path"`
-	OriginCountry       []string            `json:"origin_country"`
-	ProductionCompanies []ProductionCompany `json:"production_companies"`
-	ProductionCountries []ProductionCountry `json:"production_countries"`
-	ReleaseDate         string              `json:"release_date"`
-	Revenue             int64               `json:"revenue"`
-	Runtime             int                 `json:"runtime"`
-	SpokenLanguages     []SpokenLanguage    `json:"spoken_languages"`
-	Status              string              `json:"status"`
-	Tagline             string              `json:"tagline"`
-	Title               string              `json:"title"`
-	Video               bool                `json:"video"`
-	VoteMetrics
-	*MovieAlternativeTitlesAppend
-	*MovieChangesAppend
-	*MovieCreditsAppend
-	*MovieExternalIDsAppend
-	*MovieImagesAppend
-	*MovieKeywordsAppend
-	*MovieReleaseDatesAppend
-	*MovieVideosAppend
-	*MovieTranslationsAppend
-	*MovieRecommendationsAppend
-	*MovieSimilarAppend
-	*MovieReviewsAppend
-	*MovieListsAppend
-	*MovieWatchProvidersAppend
-}
-
-// MovieAlternativeTitlesAppend type is a struct for alternative
-// titles in append to response.
-type MovieAlternativeTitlesAppend struct {
-	AlternativeTitles *MovieAlternativeTitles `json:"alternative_titles,omitempty"`
-}
-
-// MovieChangesAppend type is a struct for changes in append to response.
-type MovieChangesAppend struct {
-	Changes *MovieChanges `json:"changes,omitempty"`
-}
-
-// MovieCreditsAppend type is a struct for credits in append to response.
-type MovieCreditsAppend struct {
-	Credits struct {
-		*MovieCredits
-	} `json:"credits,omitempty"`
-}
-
-// MovieExternalIDsAppend type is a struct for external ids in append to response.
-type MovieExternalIDsAppend struct {
-	*MovieExternalIDs `json:"external_ids,omitempty"`
-}
-
-// MovieImagesAppend type is a struct for images in append to response.
-type MovieImagesAppend struct {
-	Images *MovieImages `json:"images,omitempty"`
-}
-
-// MovieReleaseDatesAppend type is a struct for release dates in append to response.
-type MovieReleaseDatesAppend struct {
-	ReleaseDates *MovieReleaseDates `json:"release_dates,omitempty"`
-}
-
-// MovieVideosAppend type is a struct for videos in append to response.
-type MovieVideosAppend struct {
-	Videos *VideoResults `json:"videos"`
-}
-
-// MovieTranslationsAppend type is a struct for translations in append to response.
-type MovieTranslationsAppend struct {
-	Translations *MovieTranslations `json:"translations,omitempty"`
-}
-
-// MovieRecommendationsAppend type is a struct for
-// recommendations in append to response.
-type MovieRecommendationsAppend struct {
-	Recommendations *MovieRecommendations `json:"recommendations,omitempty"`
-}
-
-// MovieSimilarAppend type is a struct for similar movies in append to response.
-type MovieSimilarAppend struct {
-	Similar *MovieSimilar `json:"similar,omitempty"`
-}
-
-// MovieReviewsAppend type is a struct for reviews in append to response.
-type MovieReviewsAppend struct {
-	Reviews struct {
-		*MovieReviews
-	} `json:"reviews,omitempty"`
-}
-
-// MovieListsAppend type is a struct for lists in append to response.
-type MovieListsAppend struct {
-	Lists *MovieLists `json:"lists,omitempty"`
-}
-
-// MovieKeywordsAppend type is a struct for keywords in append to response.
-type MovieKeywordsAppend struct {
-	Keywords struct {
-		*MovieKeywords
-	} `json:"keywords,omitempty"`
-}
-
-// MovieWatchProvidersAppend type is a struct for
-// watch/providers in append to response.
-type MovieWatchProvidersAppend struct {
-	WatchProviders *WatchProviderResults `json:"watch/providers"`
-}
 
 // GetMovieDetails get the primary information about a movie.
 //
-// https://developers.themoviedb.org/3/movies
+// https://developer.themoviedb.org/reference/movie-details
 func (c *Client) GetMovieDetails(
 	id int,
 	urlOptions map[string]string,
@@ -152,14 +28,6 @@ func (c *Client) GetMovieDetails(
 	return &movieDetails, nil
 }
 
-// MovieAccountStates type is a struct for account states JSON response.
-type MovieAccountStates struct {
-	ID        int64           `json:"id"`
-	Favorite  bool            `json:"favorite"`
-	Rated     json.RawMessage `json:"rated"`
-	Watchlist bool            `json:"watchlist"`
-}
-
 // GetMovieAccountStates grab the following account states for a session:
 //
 // Movie rating.
@@ -168,7 +36,7 @@ type MovieAccountStates struct {
 //
 // If it belongs to your favourite list.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-account-states
+// https://developer.themoviedb.org/reference/movie-account-states
 func (c *Client) GetMovieAccountStates(
 	id int,
 	urlOptions map[string]string,
@@ -189,15 +57,9 @@ func (c *Client) GetMovieAccountStates(
 	return &movieAccountStates, nil
 }
 
-// MovieAlternativeTitles type is a struct for alternative titles JSON response.
-type MovieAlternativeTitles struct {
-	ID     int                `json:"id"`
-	Titles []AlternativeTitle `json:"titles"`
-}
-
 // GetMovieAlternativeTitles get all of the alternative titles for a movie.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-alternative-titles
+// https://developer.themoviedb.org/reference/movie-alternative-titles
 func (c *Client) GetMovieAlternativeTitles(
 	id int,
 	urlOptions map[string]string,
@@ -218,28 +80,13 @@ func (c *Client) GetMovieAlternativeTitles(
 	return &movieAlternativeTitles, nil
 }
 
-// MovieChanges type is a struct for changes JSON response.
-type MovieChanges struct {
-	Changes []struct {
-		Key   string `json:"key"`
-		Items []struct {
-			ID            json.RawMessage `json:"id"`
-			Action        json.RawMessage `json:"action"`
-			Time          json.RawMessage `json:"time"`
-			Iso639_1      json.RawMessage `json:"iso_639_1"`
-			Value         json.RawMessage `json:"value"`
-			OriginalValue json.RawMessage `json:"original_value"`
-		} `json:"items"`
-	} `json:"changes"`
-}
-
 // GetMovieChanges get the changes for a movie.
 //
 // By default only the last 24 hours are returned.
 // You can query up to 14 days in a single query by using
 // the start_date and end_date query parameters.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-changes
+// https://developer.themoviedb.org/reference/movie-changes
 func (c *Client) GetMovieChanges(
 	id int,
 	urlOptions map[string]string,
@@ -260,41 +107,9 @@ func (c *Client) GetMovieChanges(
 	return &movieChanges, nil
 }
 
-// MovieCredits type is a struct for credits JSON response.
-type MovieCredits struct {
-	ID   int64 `json:"id,omitempty"`
-	Cast []struct {
-		Adult              bool    `json:"adult"`
-		CastID             int64   `json:"cast_id"`
-		Character          string  `json:"character"`
-		CreditID           string  `json:"credit_id"`
-		Gender             int     `json:"gender"`
-		ID                 int64   `json:"id"`
-		KnownForDepartment string  `json:"known_for_department"`
-		Name               string  `json:"name"`
-		Order              int     `json:"order"`
-		OriginalName       string  `json:"original_name"`
-		Popularity         float32 `json:"popularity"`
-		ProfilePath        string  `json:"profile_path"`
-	} `json:"cast"`
-	Crew []struct {
-		Adult              bool    `json:"adult"`
-		CreditID           string  `json:"credit_id"`
-		Department         string  `json:"department"`
-		Gender             int     `json:"gender"`
-		ID                 int64   `json:"id"`
-		Job                string  `json:"job"`
-		KnownForDepartment string  `json:"known_for_department"`
-		Name               string  `json:"name"`
-		OriginalName       string  `json:"original_name"`
-		Popularity         float32 `json:"popularity"`
-		ProfilePath        string  `json:"profile_path"`
-	} `json:"crew"`
-}
-
 // GetMovieCredits get the cast and crew for a movie.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-credits
+// https://developer.themoviedb.org/reference/movie-credits
 func (c *Client) GetMovieCredits(
 	id int,
 	urlOptions map[string]string,
@@ -314,16 +129,6 @@ func (c *Client) GetMovieCredits(
 	return &movieCredits, nil
 }
 
-// MovieExternalIDs type is a struct for external ids JSON response.
-type MovieExternalIDs struct {
-	IMDbID      string `json:"imdb_id"`
-	FacebookID  string `json:"facebook_id"`
-	InstagramID string `json:"instagram_id"`
-	TwitterID   string `json:"twitter_id"`
-	WikiDataID  string `json:"wikidata_id,omitempty"`
-	ID          int64  `json:"id,omitempty"`
-}
-
 // GetMovieExternalIDs get the external ids for a movie.
 //
 // We currently support the following external sources.
@@ -332,7 +137,7 @@ type MovieExternalIDs struct {
 //
 // Social IDs: Facebook, Instagram and Twitter.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-external-ids
+// https://developer.themoviedb.org/reference/movie-external-ids
 func (c *Client) GetMovieExternalIDs(
 	id int,
 	urlOptions map[string]string,
@@ -353,20 +158,6 @@ func (c *Client) GetMovieExternalIDs(
 	return &movieExternalIDs, nil
 }
 
-// MovieImage type is a struct for a single image.
-type MovieImage struct {
-	ImageBase
-	Iso639_1 string `json:"iso_639_1"`
-}
-
-// MovieImages type is a struct for images JSON response.
-type MovieImages struct {
-	ID        int64        `json:"id,omitempty"`
-	Backdrops []MovieImage `json:"backdrops"`
-	Logos     []MovieImage `json:"logos"`
-	Posters   []MovieImage `json:"posters"`
-}
-
 // GetMovieImages get the images that belong to a movie.
 //
 // Querying images with a language parameter will filter the results.
@@ -374,7 +165,7 @@ type MovieImages struct {
 // you can use the include_image_language parameter. This should be a comma
 // separated value like so: include_image_language=en,null.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-images
+// https://developer.themoviedb.org/reference/movie-images
 func (c *Client) GetMovieImages(
 	id int,
 	urlOptions map[string]string,
@@ -395,18 +186,9 @@ func (c *Client) GetMovieImages(
 	return &movieImages, nil
 }
 
-// MovieKeywords type is a struct for keywords JSON response.
-type MovieKeywords struct {
-	ID       int64 `json:"id,omitempty"`
-	Keywords []struct {
-		ID   int64  `json:"id"`
-		Name string `json:"name"`
-	} `json:"keywords"`
-}
-
 // GetMovieKeywords get the keywords that have been added to a movie.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-keywords
+// https://developer.themoviedb.org/reference/movie-keywords
 func (c *Client) GetMovieKeywords(id int) (*MovieKeywords, error) {
 	tmdbURL := fmt.Sprintf(
 		"%s%s%d/keywords?api_key=%s",
@@ -422,15 +204,9 @@ func (c *Client) GetMovieKeywords(id int) (*MovieKeywords, error) {
 	return &movieKeywords, nil
 }
 
-// MovieReleaseDates type is a struct for release dates JSON response.
-type MovieReleaseDates struct {
-	ID int64 `json:"id,omitempty"`
-	*MovieReleaseDatesResults
-}
-
 // GetMovieReleaseDates get the release date along with the certification for a movie.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-release-dates
+// https://developer.themoviedb.org/reference/movie-release-dates
 func (c *Client) GetMovieReleaseDates(
 	id int,
 ) (*MovieReleaseDates, error) {
@@ -450,7 +226,7 @@ func (c *Client) GetMovieReleaseDates(
 
 // GetMovieVideos get the videos that have been added to a movie.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-videos
+// https://developer.themoviedb.org/reference/movie-videos
 func (c *Client) GetMovieVideos(
 	id int,
 	urlOptions map[string]string,
@@ -473,7 +249,7 @@ func (c *Client) GetMovieVideos(
 
 // GetMovieWatchProviders get a list of the availabilities per country by provider for a movie.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-watch-providers
+// https://developer.themoviedb.org/reference/movie-watch-providers
 func (c *Client) GetMovieWatchProviders(
 	id int,
 	urlOptions map[string]string,
@@ -494,15 +270,9 @@ func (c *Client) GetMovieWatchProviders(
 	return &movieWatchProviders, nil
 }
 
-// MovieTranslations type is a struct for translations JSON response.
-type MovieTranslations struct {
-	ID           int64         `json:"id,omitempty"`
-	Translations []Translation `json:"translations"`
-}
-
 // GetMovieTranslations get a list of translations that have been created for a movie.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-translations
+// https://developer.themoviedb.org/reference/movie-translations
 func (c *Client) GetMovieTranslations(
 	id int,
 	urlOptions map[string]string,
@@ -523,15 +293,9 @@ func (c *Client) GetMovieTranslations(
 	return &movieTranslations, nil
 }
 
-// MovieRecommendations type is a struct for recommendations JSON response.
-type MovieRecommendations struct {
-	*MovieRecommendationsResults
-	PaginatedResultsMeta
-}
-
 // GetMovieRecommendations get a list of recommended movies for a movie.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-recommendations
+// https://developer.themoviedb.org/reference/movie-recommendations
 func (c *Client) GetMovieRecommendations(
 	id int,
 	urlOptions map[string]string,
@@ -552,17 +316,12 @@ func (c *Client) GetMovieRecommendations(
 	return &movieRecommendations, nil
 }
 
-// MovieSimilar type is a struct for similar movies JSON response.
-type MovieSimilar struct {
-	*MovieRecommendations
-}
-
 // GetMovieSimilar get a list of similar movies.
 //
 // This is not the same as the "Recommendation" system you see on the website.
 // These items are assembled by looking at keywords and genres.
 //
-// https://developers.themoviedb.org/3/movies/get-similar-movies
+// https://developer.themoviedb.org/reference/movie-similar
 func (c *Client) GetMovieSimilar(
 	id int,
 	urlOptions map[string]string,
@@ -583,16 +342,9 @@ func (c *Client) GetMovieSimilar(
 	return &movieSimilar, nil
 }
 
-// MovieReviews type is a struct for reviews JSON response.
-type MovieReviews struct {
-	ID int64 `json:"id,omitempty"`
-	*MovieReviewsResults
-	PaginatedResultsMeta
-}
-
 // GetMovieReviews get the user reviews for a movie.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-reviews
+// https://developer.themoviedb.org/reference/movie-reviews
 func (c *Client) GetMovieReviews(
 	id int,
 	urlOptions map[string]string,
@@ -613,16 +365,9 @@ func (c *Client) GetMovieReviews(
 	return &movieReviews, nil
 }
 
-// MovieLists type is a struct for lists JSON response.
-type MovieLists struct {
-	ID int64 `json:"id"`
-	*MovieListsResults
-	PaginatedResultsMeta
-}
-
 // GetMovieLists get a list of lists that this movie belongs to.
 //
-// https://developers.themoviedb.org/3/movies/get-movie-lists
+// https://developer.themoviedb.org/reference/movie-lists
 func (c *Client) GetMovieLists(
 	id int,
 	urlOptions map[string]string,
@@ -643,16 +388,11 @@ func (c *Client) GetMovieLists(
 	return &movieLists, nil
 }
 
-// MovieLatest type is a struct for latest JSON response.
-type MovieLatest struct {
-	*MovieDetails
-}
-
 // GetMovieLatest get the most newly created movie.
 //
 // This is a live response and will continuously change.
 //
-// https://developers.themoviedb.org/3/movies/get-latest-movie
+// https://developer.themoviedb.org/reference/movie-latest-id
 func (c *Client) GetMovieLatest(
 	urlOptions map[string]string,
 ) (*MovieLatest, error) {
@@ -671,16 +411,6 @@ func (c *Client) GetMovieLatest(
 	return &movieLastest, nil
 }
 
-// MovieNowPlaying type is a struct for now playing JSON response.
-type MovieNowPlaying struct {
-	*MovieNowPlayingResults
-	Dates struct {
-		Maximum string `json:"maximum"`
-		Minimum string `json:"minimum"`
-	} `json:"dates"`
-	PaginatedResultsMeta
-}
-
 // GetMovieNowPlaying get a list of movies in theatres.
 //
 // This is a release type query that looks for all movies that
@@ -690,7 +420,7 @@ type MovieNowPlaying struct {
 // the search to only look for theatrical release dates within the
 // specified country.
 //
-// https://developers.themoviedb.org/3/movies/get-now-playing
+// https://developer.themoviedb.org/reference/movie-now-playing-list
 func (c *Client) GetMovieNowPlaying(
 	urlOptions map[string]string,
 ) (*MovieNowPlaying, error) {
@@ -709,17 +439,11 @@ func (c *Client) GetMovieNowPlaying(
 	return &movieNowPlaying, nil
 }
 
-// MoviePopular type is a struct for popular JSON response.
-type MoviePopular struct {
-	*MoviePopularResults
-	PaginatedResultsMeta
-}
-
 // GetMoviePopular get a list of the current popular movies on TMDb.
 //
 // This list updates daily.
 //
-// https://developers.themoviedb.org/3/movies/get-popular-movies
+// https://developer.themoviedb.org/reference/movie-popular-list
 func (c *Client) GetMoviePopular(
 	urlOptions map[string]string,
 ) (*MoviePopular, error) {
@@ -738,14 +462,9 @@ func (c *Client) GetMoviePopular(
 	return &moviePopular, nil
 }
 
-// MovieTopRated type is a struct for top rated JSON response.
-type MovieTopRated struct {
-	*MoviePopular
-}
-
 // GetMovieTopRated get the top rated movies on TMDb.
 //
-// https://developers.themoviedb.org/3/movies/get-top-rated-movies
+// https://developer.themoviedb.org/reference/movie-top-rated-list
 func (c *Client) GetMovieTopRated(
 	urlOptions map[string]string,
 ) (*MovieTopRated, error) {
@@ -764,11 +483,6 @@ func (c *Client) GetMovieTopRated(
 	return &movieTopRated, nil
 }
 
-// MovieUpcoming type is a struct for upcoming JSON response.
-type MovieUpcoming struct {
-	*MovieNowPlaying
-}
-
 // GetMovieUpcoming get a list of upcoming movies in theatres.
 //
 // This is a release type query that looks for all movies that
@@ -778,7 +492,7 @@ type MovieUpcoming struct {
 // the search to only look for theatrical release dates within
 // the specified country.
 //
-// https://developers.themoviedb.org/3/movies/get-upcoming
+// https://developer.themoviedb.org/reference/movie-upcoming-list
 func (c *Client) GetMovieUpcoming(
 	urlOptions map[string]string,
 ) (*MovieUpcoming, error) {
@@ -802,9 +516,9 @@ func (c *Client) GetMovieUpcoming(
 // A valid session or guest session ID is required.
 //
 // You can read more about how this works:
-// https://developers.themoviedb.org/3/authentication/how-do-i-generate-a-session-id
+// https://developer.themoviedb.org/reference/authentication-how-do-i-generate-a-session-id
 //
-// https://developers.themoviedb.org/3/movies/rate-movie
+// https://developer.themoviedb.org/reference/movie-add-rating
 func (c *Client) PostMovieRating(
 	id int,
 	rating float32,
@@ -840,9 +554,9 @@ func (c *Client) PostMovieRating(
 // A valid session or guest session ID is required.
 //
 // You can read more about how this works:
-// https://developers.themoviedb.org/3/authentication/how-do-i-generate-a-session-id
+// https://developer.themoviedb.org/reference/authentication-how-do-i-generate-a-session-id
 //
-// https://developers.themoviedb.org/3/movies/delete-movie-rating
+// https://developer.themoviedb.org/reference/movie-delete-rating
 func (c *Client) DeleteMovieRating(
 	id int,
 	urlOptions map[string]string,
