+++ movies_struct.go
@@ -1,0 +1,287 @@
+package tmdb
+
+import json "github.com/goccy/go-json"
+
+// MovieDetails type is a struct for movie details JSON response.
+type MovieDetails struct {
+	Adult               bool                `json:"adult"`
+	BackdropPath        string              `json:"backdrop_path"`
+	BelongsToCollection BelongsToCollection `json:"belongs_to_collection"`
+	Budget              int64               `json:"budget"`
+	Genres              []Genre             `json:"genres"`
+	Homepage            string              `json:"homepage"`
+	ID                  int64               `json:"id"`
+	IMDbID              string              `json:"imdb_id"`
+	OriginalLanguage    string              `json:"original_language"`
+	OriginalTitle       string              `json:"original_title"`
+	Overview            string              `json:"overview"`
+	Popularity          float32             `json:"popularity"`
+	PosterPath          string              `json:"poster_path"`
+	OriginCountry       []string            `json:"origin_country"`
+	ProductionCompanies []ProductionCompany `json:"production_companies"`
+	ProductionCountries []ProductionCountry `json:"production_countries"`
+	ReleaseDate         string              `json:"release_date"`
+	Revenue             int64               `json:"revenue"`
+	Runtime             int                 `json:"runtime"`
+	SpokenLanguages     []SpokenLanguage    `json:"spoken_languages"`
+	Status              string              `json:"status"`
+	Tagline             string              `json:"tagline"`
+	Title               string              `json:"title"`
+	Video               bool                `json:"video"`
+	VoteMetrics
+	*MovieAlternativeTitlesAppend
+	*MovieChangesAppend
+	*MovieCreditsAppend
+	*MovieExternalIDsAppend
+	*MovieImagesAppend
+	*MovieKeywordsAppend
+	*MovieReleaseDatesAppend
+	*MovieVideosAppend
+	*MovieTranslationsAppend
+	*MovieRecommendationsAppend
+	*MovieSimilarAppend
+	*MovieReviewsAppend
+	*MovieListsAppend
+	*MovieWatchProvidersAppend
+}
+
+// MovieAlternativeTitlesAppend type is a struct for alternative
+// titles in append to response.
+type MovieAlternativeTitlesAppend struct {
+	AlternativeTitles *MovieAlternativeTitles `json:"alternative_titles,omitempty"`
+}
+
+// MovieChangesAppend type is a struct for changes in append to response.
+type MovieChangesAppend struct {
+	Changes *MovieChanges `json:"changes,omitempty"`
+}
+
+// MovieCreditsAppend type is a struct for credits in append to response.
+type MovieCreditsAppend struct {
+	Credits struct {
+		*MovieCredits
+	} `json:"credits,omitempty"`
+}
+
+// MovieExternalIDsAppend type is a struct for external ids in append to response.
+type MovieExternalIDsAppend struct {
+	*MovieExternalIDs `json:"external_ids,omitempty"`
+}
+
+// MovieImagesAppend type is a struct for images in append to response.
+type MovieImagesAppend struct {
+	Images *MovieImages `json:"images,omitempty"`
+}
+
+// MovieReleaseDatesAppend type is a struct for release dates in append to response.
+type MovieReleaseDatesAppend struct {
+	ReleaseDates *MovieReleaseDates `json:"release_dates,omitempty"`
+}
+
+// MovieVideosAppend type is a struct for videos in append to response.
+type MovieVideosAppend struct {
+	Videos *VideoResults `json:"videos"`
+}
+
+// MovieTranslationsAppend type is a struct for translations in append to response.
+type MovieTranslationsAppend struct {
+	Translations *MovieTranslations `json:"translations,omitempty"`
+}
+
+// MovieRecommendationsAppend type is a struct for
+// recommendations in append to response.
+type MovieRecommendationsAppend struct {
+	Recommendations *MovieRecommendations `json:"recommendations,omitempty"`
+}
+
+// MovieSimilarAppend type is a struct for similar movies in append to response.
+type MovieSimilarAppend struct {
+	Similar *MovieSimilar `json:"similar,omitempty"`
+}
+
+// MovieReviewsAppend type is a struct for reviews in append to response.
+type MovieReviewsAppend struct {
+	Reviews struct {
+		*MovieReviews
+	} `json:"reviews,omitempty"`
+}
+
+// MovieListsAppend type is a struct for lists in append to response.
+type MovieListsAppend struct {
+	Lists *MovieLists `json:"lists,omitempty"`
+}
+
+// MovieKeywordsAppend type is a struct for keywords in append to response.
+type MovieKeywordsAppend struct {
+	Keywords struct {
+		*MovieKeywords
+	} `json:"keywords,omitempty"`
+}
+
+// MovieWatchProvidersAppend type is a struct for
+// watch/providers in append to response.
+type MovieWatchProvidersAppend struct {
+	WatchProviders *WatchProviderResults `json:"watch/providers"`
+}
+
+// MovieAccountStates type is a struct for account states JSON response.
+type MovieAccountStates struct {
+	ID        int64           `json:"id"`
+	Favorite  bool            `json:"favorite"`
+	Rated     json.RawMessage `json:"rated"`
+	Watchlist bool            `json:"watchlist"`
+}
+
+// MovieAlternativeTitles type is a struct for alternative titles JSON response.
+type MovieAlternativeTitles struct {
+	ID     int                `json:"id"`
+	Titles []AlternativeTitle `json:"titles"`
+}
+
+// MovieChanges type is a struct for changes JSON response.
+type MovieChanges struct {
+	Changes []struct {
+		Key   string `json:"key"`
+		Items []struct {
+			ID            json.RawMessage `json:"id"`
+			Action        json.RawMessage `json:"action"`
+			Time          json.RawMessage `json:"time"`
+			Iso639_1      json.RawMessage `json:"iso_639_1"`
+			Value         json.RawMessage `json:"value"`
+			OriginalValue json.RawMessage `json:"original_value"`
+		} `json:"items"`
+	} `json:"changes"`
+}
+
+// MovieCredits type is a struct for credits JSON response.
+type MovieCredits struct {
+	ID   int64 `json:"id,omitempty"`
+	Cast []struct {
+		Adult              bool    `json:"adult"`
+		CastID             int64   `json:"cast_id"`
+		Character          string  `json:"character"`
+		CreditID           string  `json:"credit_id"`
+		Gender             int     `json:"gender"`
+		ID                 int64   `json:"id"`
+		KnownForDepartment string  `json:"known_for_department"`
+		Name               string  `json:"name"`
+		Order              int     `json:"order"`
+		OriginalName       string  `json:"original_name"`
+		Popularity         float32 `json:"popularity"`
+		ProfilePath        string  `json:"profile_path"`
+	} `json:"cast"`
+	Crew []struct {
+		Adult              bool    `json:"adult"`
+		CreditID           string  `json:"credit_id"`
+		Department         string  `json:"department"`
+		Gender             int     `json:"gender"`
+		ID                 int64   `json:"id"`
+		Job                string  `json:"job"`
+		KnownForDepartment string  `json:"known_for_department"`
+		Name               string  `json:"name"`
+		OriginalName       string  `json:"original_name"`
+		Popularity         float32 `json:"popularity"`
+		ProfilePath        string  `json:"profile_path"`
+	} `json:"crew"`
+}
+
+// MovieExternalIDs type is a struct for external ids JSON response.
+type MovieExternalIDs struct {
+	IMDbID      string `json:"imdb_id"`
+	FacebookID  string `json:"facebook_id"`
+	InstagramID string `json:"instagram_id"`
+	TwitterID   string `json:"twitter_id"`
+	WikiDataID  string `json:"wikidata_id,omitempty"`
+	ID          int64  `json:"id,omitempty"`
+}
+
+// MovieImage type is a struct for a single image.
+type MovieImage struct {
+	ImageBase
+	Iso639_1 string `json:"iso_639_1"`
+}
+
+// MovieImages type is a struct for images JSON response.
+type MovieImages struct {
+	ID        int64        `json:"id,omitempty"`
+	Backdrops []MovieImage `json:"backdrops"`
+	Logos     []MovieImage `json:"logos"`
+	Posters   []MovieImage `json:"posters"`
+}
+
+// MovieKeywords type is a struct for keywords JSON response.
+type MovieKeywords struct {
+	ID       int64 `json:"id,omitempty"`
+	Keywords []struct {
+		ID   int64  `json:"id"`
+		Name string `json:"name"`
+	} `json:"keywords"`
+}
+
+// MovieReleaseDates type is a struct for release dates JSON response.
+type MovieReleaseDates struct {
+	ID int64 `json:"id,omitempty"`
+	*MovieReleaseDatesResults
+}
+
+// MovieTranslations type is a struct for translations JSON response.
+type MovieTranslations struct {
+	ID           int64         `json:"id,omitempty"`
+	Translations []Translation `json:"translations"`
+}
+
+// MovieRecommendations type is a struct for recommendations JSON response.
+type MovieRecommendations struct {
+	*MovieRecommendationsResults
+	PaginatedResultsMeta
+}
+
+// MovieSimilar type is a struct for similar movies JSON response.
+type MovieSimilar struct {
+	*MovieRecommendations
+}
+
+// MovieReviews type is a struct for reviews JSON response.
+type MovieReviews struct {
+	ID int64 `json:"id,omitempty"`
+	*MovieReviewsResults
+	PaginatedResultsMeta
+}
+
+// MovieLists type is a struct for lists JSON response.
+type MovieLists struct {
+	ID int64 `json:"id"`
+	*MovieListsResults
+	PaginatedResultsMeta
+}
+
+// MovieLatest type is a struct for latest JSON response.
+type MovieLatest struct {
+	*MovieDetails
+}
+
+// MovieNowPlaying type is a struct for now playing JSON response.
+type MovieNowPlaying struct {
+	*MovieNowPlayingResults
+	Dates struct {
+		Maximum string `json:"maximum"`
+		Minimum string `json:"minimum"`
+	} `json:"dates"`
+	PaginatedResultsMeta
+}
+
+// MoviePopular type is a struct for popular JSON response.
+type MoviePopular struct {
+	*MoviePopularResults
+	PaginatedResultsMeta
+}
+
+// MovieTopRated type is a struct for top rated JSON response.
+type MovieTopRated struct {
+	*MoviePopular
+}
+
+// MovieUpcoming type is a struct for upcoming JSON response.
+type MovieUpcoming struct {
+	*MovieNowPlaying
+}
