+++ networks.go
@@ -2,19 +2,9 @@ package tmdb
 
 import "fmt"
 
-// NetworkDetails type is a struct for details JSON response.
-type NetworkDetails struct {
-	Headquarters  string `json:"headquarters"`
-	Homepage      string `json:"homepage"`
-	ID            int64  `json:"id"`
-	LogoPath      string `json:"logo_path"`
-	Name          string `json:"name"`
-	OriginCountry string `json:"origin_country"`
-}
-
 // GetNetworkDetails get the details of a network.
 //
-// https://developers.themoviedb.org/3/networks/get-network-details
+// https://developer.themoviedb.org/reference/network-details
 func (c *Client) GetNetworkDetails(
 	id int,
 ) (*NetworkDetails, error) {
@@ -32,18 +22,9 @@ func (c *Client) GetNetworkDetails(
 	return &networkDetails, nil
 }
 
-// NetworkAlternativeNames type is a struct for alternative names JSON response.
-type NetworkAlternativeNames struct {
-	ID      int64 `json:"id"`
-	Results []struct {
-		Name string `json:"name"`
-		Type string `json:"type"`
-	} `json:"results"`
-}
-
 // GetNetworkAlternativeNames get the alternative names of a network.
 //
-// https://developers.themoviedb.org/3/networks/get-network-alternative-names
+// https://developer.themoviedb.org/reference/details-copy
 func (c *Client) GetNetworkAlternativeNames(
 	id int,
 ) (*NetworkAlternativeNames, error) {
@@ -61,19 +42,6 @@ func (c *Client) GetNetworkAlternativeNames(
 	return &networkAltenativeNames, nil
 }
 
-// NetworkImage type is a struct for a single image.
-type NetworkImage struct {
-	ImageBase
-	ID       string `json:"id"`
-	FileType string `json:"file_type"`
-}
-
-// NetworkImages type is a struct for images JSON response.
-type NetworkImages struct {
-	ID    int64          `json:"id"`
-	Logos []NetworkImage `json:"logos"`
-}
-
 // GetNetworkImages get the TV network logos by id.
 //
 // There are two image formats that are supported for networks,
@@ -84,7 +52,7 @@ type NetworkImages struct {
 // An SVG can be scaled properly beyond those dimensions if you
 // call them as a PNG.
 //
-// https://developers.themoviedb.org/3/networks/get-network-images
+// https://developer.themoviedb.org/reference/alternative-names-copy
 func (c *Client) GetNetworkImages(
 	id int,
 ) (*NetworkImages, error) {
