+++ trending_struct.go
@@ -1,0 +1,7 @@
+package tmdb
+
+// Trending type is a struct for trending JSON response.
+type Trending struct {
+	*TrendingResults
+	PaginatedResultsMeta
+}
